fix(connect): don't import from browser file
