+++ packages/connect/src/utils/assetUtils.ts
@@ -1,3 +1,5 @@
+import fetch from 'cross-fetch';
+
 import { firmwareAssets } from '@trezor/connect-data';
 import connectDataCoinsEth from '@trezor/connect-data/files/coins-eth.json';
 import connectDataCoins from '@trezor/connect-data/files/coins.json';
@@ -6,6 +8,8 @@ import type { DeviceModelInternal, FirmwareRelease } from '@trezor/device-utils'
 import { FirmwareType } from '@trezor/device-utils';
 import { versionUtils } from '@trezor/utils';
 import type { VersionArray } from '@trezor/utils/src/versionUtils';
+
+import type { HttpRequestOptions, HttpRequestReturnType, HttpRequestType } from './assetsTypes';
 
 export class HttpRequestError extends Error {
     response: Response;
@@ -72,4 +76,31 @@ export const tryLocalAssetRequire = (url: string): unknown => {
     }
 
     return null;
+};
+
+/**
+ * Http request wrapper for Suite Web & Desktop to handle various response states in a unified way.
+ */
+export const httpRequest = async <T extends HttpRequestType>(
+    url: string,
+    type: T = 'text' as T,
+    options?: HttpRequestOptions,
+): Promise<HttpRequestReturnType<T>> => {
+    const init: RequestInit = { ...options, credentials: 'same-origin' };
+
+    const response = await fetch(url, init);
+    if (response.ok) {
+        if (type === 'json') {
+            const txt = await response.text();
+
+            return JSON.parse(txt) as HttpRequestReturnType<T>;
+        }
+        if (type === 'binary') {
+            return response.arrayBuffer() as Promise<HttpRequestReturnType<T>>;
+        }
+
+        return response.text() as Promise<HttpRequestReturnType<T>>;
+    }
+
+    throw new HttpRequestError(response);
 };
