+++ packages/connect/src/utils/assets.ts
@@ -1,8 +1,7 @@
 import fetch from 'cross-fetch';
 import { promises as fs } from 'fs';
 
-import { tryLocalAssetRequire } from './assetUtils';
-import { httpRequest as browserHttpRequest } from './assets.browser';
+import { httpRequest as browserHttpRequest, tryLocalAssetRequire } from './assetUtils';
 import type { HttpRequestOptions, HttpRequestReturnType, HttpRequestType } from './assetsTypes';
 
 if (global && typeof global.fetch !== 'function') {
