+++ packages/transport/src/transports/nodeusb.native.ts
@@ -2,7 +2,7 @@ import { AbstractTransport, type AbstractTransportParams } from './abstract';
 import { WRONG_ENVIRONMENT } from '../errors';
 import { empty, emptySync } from '../utils/resultEmpty';
 
-// this class loads in browser environment only in case of accidental use of NodeUsbTransport
+// this class loads in React Native environment only in case of accidental use of NodeUsbTransport
 
 export class NodeUsbTransport extends AbstractTransport {
     public name = 'NodeUsbTransport' as const;
